+++ korio/src/commonMain/kotlin/com/soywiz/korio/async/withOptTimeout.kt
@@ -2,7 +2,7 @@ package com.soywiz.korio.async
 
 import kotlinx.coroutines.*
 
-suspend fun <T> withOptTimeout(ms: Int?, name: String = "timeout", callback: suspend () -> T): T {
+suspend fun <T> withOptTimeout(ms: Long?, name: String = "timeout", callback: suspend () -> T): T {
 	if (ms == null) return callback()
 	return withTimeout(ms) { callback() }
 }
